snp: Support VLEK
EC2 SNP instances report an all zero chipid, and it is not possible to
fetch a VCEK for them. Instead they support a "VLEK" that can be fetched
using the SNP_GET_EXT_REPORT ioctl. Add support for retrieving
certificates with the extended guest report and fetching the VLEK chain,
as VLEK uses a different intermediate certificate.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>